+++ packages/dart/lib/src/format.dart
@@ -64,7 +64,7 @@ int safeParseInt(dynamic value, {int? fallback, int? radix}) {
 /// Attempts to parse without returning an error, even if the argument value is of the wrong type or in `JSON` format.
 /// If parsing fails, it will be replaced with the object set in `fallback`.
 /// The default value for `fallback` is an empty object.
-dynamic safeJSONParse(dynamic jsonString, [dynamic fallback = const {}]) {
+dynamic safeJSONParse(dynamic jsonString, {dynamic fallback = const {}}) {
   if (jsonString == null) {
     return fallback;
   }