+++ packages/dart/CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 1.1.7 (2025--)
 
 - **BREAKING CHANGES**: `getFileSize` has renamed to `fileSizeFormat`
-- **BREAKING CHANGES**: `safeJSONParse`: fallback parameters has changed to named parameter
+- **BREAKING CHANGES**: `safeJSONParse`: 'fallback' parameters has changed to named parameter
+- **BREAKING CHANGES**: `objToArray`: 'recursive' parameters has changed to named parameter
 - Add `getFileName` and `getFileExtension` methods
 - Add `isFileExists` method